Created dashboard
Dashboard logic:
General dashboard file for all the users. Through retrieving the user id in the auth, we can collect the role id from the user. We can then put the div (square) into an if function that will check the role id with the id in roles.php file with the role names + ids